% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CryptDriftR.R
\docType{package}
\name{CryptDriftR-package}
\alias{CryptDriftR}
\alias{CryptDriftR-package}
\title{Simulate, analyse and plot crypt clonal data}
\description{
A set of functions are provided to simulate neutral drift, biased drift and continuous labelling.
Functions are also provided to infer drift parameters from crypt count data.  
The models are explained in detail in "Continuous clonal labeling reveals small numbers of functional stem cells in intestinal crypts and adenomas." 
Kozar, Morrissey et. al. 2013 Cell Stem Cell, as well as "Defining stem cell dynamics in models of intestinal tumor initiation." 
Vermeulen, Morrissey, et. al. Science 2013.
}
