% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BiasedDrift.R
\name{fitBiasedDrift}
\alias{fitBiasedDrift}
\title{Fit crypt biased stem cell drift model.}
\usage{
fitBiasedDrift(x, time_interval, WT_params, max_iter = 80000,
  n_par_chains = 4, burn_in = 5000, thin = 50)
}
\arguments{
\item{x}{Clone size counts. With a column per day and a row per size clone size.}

\item{time_interval}{vector of time point values. Should have same number of elements as columns in x.}

\item{WT_params}{List with wild type neutral drift parameters (N, lambda and tau).}

\item{max_iter}{number of mcmc iterations to run.}

\item{n_par_chains}{number of parallel chains (and cpus used) to run.}

\item{burn_in}{mcmc burn in}

\item{thin}{mcmc thining}
}
\value{
Will retrun a list that can be used directly with the plotting functions. The list contains an mcmc matrix, the data and time vector used to fit, as well 
as the MAP estimate of the parameters.
}
\description{
\code{fitBiasedDrift} will run an mcmc algorithm to infer the parameters from the biased drift model.
}
\examples{
time_points = c(4, 7, 10, 14, 21)
WT_params = list(lambda = 0.1, N  = 5, tau = 1)
x = simulateBiasedDriftData(WT_params$lambda, WT_params$N, WT_params$tau, Pr = 0.7, time_points, 8, 300)

fit_out = fitBiasedDrift(x, time_points, WT_params)
plotsConvergence_Biased(fit_out)
plotsBiasDrift_Fit(fit_out)

}
