% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_NeutralDrift.R
\name{fitNeutralDrift}
\alias{fitNeutralDrift}
\title{Fit crypt neutral drift model.}
\usage{
fitNeutralDrift(x, time_interval, max_iter = 40000, n_par_chains = 4,
  burn_in = 5000, thin = 20)
}
\arguments{
\item{x}{Clone size counts. With a column per day and a row per size clone size.}

\item{time_interval}{vector of time point values. Should have same number of elements as columns in x.}

\item{max_iter}{number of mcmc iterations to run.}

\item{n_par_chains}{number of parallel chains (and cpus used) to run.}

\item{burn_in}{mcmc burn in}

\item{thin}{mcmc thining}
}
\value{
Will retrun a list that can be used directly with the plotting functions. The list contains an mcmc matrix, the data and time vector used to fit
}
\description{
\code{fitNeutralDrift}
}
\examples{
time_points = c(4, 7, 10, 14, 21)
x = simulateNeutralDriftData(0.1, 5, 2, time_points, 8, 300)
fit_out = fitNeutralDrift(x, time_points)
plotsConvergence_Neutral(fit_out)
plotsNeutralDrift_Fit(fit_out)

}
