% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BiasedDrift.R
\name{plotPosterior_Pr}
\alias{plotPosterior_Pr}
\title{MCMC inference plots for biased drift model}
\usage{
plotPosterior_Pr(fit_vals)
}
\arguments{
\item{fit_vals}{list produced by fitBiasedDrift containing amongst other things the MCMC chains.}
}
\value{
Return ggplot object with posterior plots.
}
\description{
\code{plotPosterior_Pr} will plot posterior plots using the MCMC chains.
}
\examples{
time_points = c(4, 7, 10, 14, 21)
WT_params = list(lambda = 0.1, N  = 5, tau = 1)
x = simulateBiasedDriftData(WT_params$lambda, WT_params$N, WT_params$tau, Pr = 0.7, time_points, 8, 300)

fit_out = fitBiasedDrift(x, WT_params, time_points)
plotsConvergence_Biased(fit_out)
plotPosterior_Pr(fit_out)

}
