% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BiasedDrift.R
\name{plotsBiasDrift_Fit}
\alias{plotsBiasDrift_Fit}
\title{Plot data with biased drift model fit}
\usage{
plotsBiasDrift_Fit(fit_vals, max_x = 100)
}
\arguments{
\item{fit_vals}{list produced by fitBiasedDrift containing amongst other things the MCMC chains.}

\item{max_x}{Choose maximum value for the x axis (time)}
}
\value{
Return ggplot object with posterior plots.
}
\description{
\code{plotsBiasDrift_Fit} will plot  data and fit of biased drift model.
}
\examples{
time_points = c(4, 7, 10, 14, 21)
WT_params = list(lambda = 0.1, N  = 5, tau = 1)
x = simulateBiasedDriftData(WT_params$lambda, WT_params$N, WT_params$tau, Pr = 0.7, time_points, 8, 300)

fit_out = fitBiasedDrift(x, WT_params, time_points)
plotsConvergence_Biased(fit_out)
plotsBiasDrift_Fit(fit_out)

}
