% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_NeutralDrift.R
\name{plotsNeutralDrift_Fit}
\alias{plotsNeutralDrift_Fit}
\title{Plot data with neutral drift model fit}
\usage{
plotsNeutralDrift_Fit(fit_vals, max_x = 100)
}
\arguments{
\item{fit_vals}{list produced by fitNeutralDrift containing amongst other things the MCMC chains.}

\item{max_x}{Choose maximum value for the x axis (time)}
}
\value{
Return ggplot object with posterior plots.
}
\description{
\code{plotsNeutralDrift_Fit} will plot  data and fit of neutral drift model.
}
\examples{
time_points = c(4, 7, 10, 14, 21)
x = simulateNeutralDriftData(0.1, 5, 2, time_points, 8, 300)
fit_out = fitNeutralDrift(x, time_points)
plotsConvergence_Neutral(fit_out)
plotsNeutralDrift_Fit(fit_out)

}
