% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BiasedDrift.R
\name{simulateBiasedDriftData}
\alias{simulateBiasedDriftData}
\title{Simulate data from biased drift model for testing inference.}
\usage{
simulateBiasedDriftData(lambda, Ns, tau, Pr, time_points, size_measure,
  num_crypts)
}
\arguments{
\item{lambda}{Stem cell replacement rate.}

\item{Ns}{Number of functional stem cells per crypt.}

\item{tau}{Time delay until drift starts.}

\item{Pr}{Bias in stem cell replacement Pr = 0.5 is neutral.}

\item{time_points}{vector of time point values.}

\item{size_measure}{Fractions in which the crypts are measured in (e.g. 8ths).}

\item{num_crypts}{Number of crypts measured per time point.}
}
\value{
Will retrun simulated crypt clone count data.
}
\description{
\code{simulateBiasedDriftData} will use the analytical biased stem cell drift model solutions to similate count data.
}
\examples{
time_points = c(4, 7, 10, 14, 21)
WT_params = list(lambda = 0.1, N  = 5, tau = 1)
x = simulateBiasedDriftData(WT_params$lambda, WT_params$N, WT_params$tau, Pr = 0.7, time_points, 8, 300)

}
