% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_NeutralDrift.R
\name{simulateNeutralDriftData}
\alias{simulateNeutralDriftData}
\title{Simulate data from neutral drift model for testing inference.}
\usage{
simulateNeutralDriftData(lambda, Ns, tau, time_points, size_measure, num_crypts)
}
\arguments{
\item{lambda}{Stem cell replacement rate.}

\item{Ns}{Number of functional stem cells per crypt.}

\item{tau}{Time delay until drift starts.}

\item{time_points}{vector of time point values.}

\item{size_measure}{Fractions in which the crypts are measured in (e.g. 8ths).}

\item{num_crypts}{Number of crypts measured per time point.}
}
\value{
Will retrun simulated crypt clone count data.
}
\description{
\code{simulateNeutralDriftData} will use the analytical neutral drift model solutions to similate count data.
}
\examples{
time_points = c(4, 7, 10, 14, 21)
x = simulateNeutralDriftData(0.1, 5, 2, time_points, 8, 300)

}
