% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriftFunctions.R
\name{th_contLabelling}
\alias{th_contLabelling}
\title{Analytical approximation to continuous labelling model.}
\usage{
th_contLabelling(mu, lambda, Ns, time_points)
}
\arguments{
\item{mu}{Stem cell replacement rate.}

\item{lambda}{Stem cell replacement rate.}

\item{Ns}{Number of functional stem cells per crypt.}

\item{time_points}{vector of time point values.}
}
\value{
a matrix with two rows, the frequency of partial clone cypts and the frequency of monoclonal crypts.
}
\description{
\code{th_contLabelling}
}
\examples{
x = th_contLabelling(mu = 1.1e-4, lambda = 0.1, Ns = 5, time_points = 101:201)

}
