% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriftFunctions.R
\name{th_contLabelling_slope_and_partials}
\alias{th_contLabelling_slope_and_partials}
\title{Analytical approximation to continuous labelling model, also allows bias. 
\code{th_contLabelling}}
\usage{
th_contLabelling_slope_and_partials(mu, lambda, Ns, Pr = 0.5,
  alpha_beta = NULL)
}
\arguments{
\item{mu}{Stem cell replacement rate.}

\item{lambda}{Stem cell replacement rate.}

\item{Ns}{Number of functional stem cells per crypt.}
}
\value{
A list with the slope of the monoclonal accumulation, the intercept of the monoclonal accumulation and the level of the partials.
}
\description{
The function returns only the slope of the monoclonal crypts and the partial clone levels.
}
\examples{
x = th_contLabelling_slope_and_partials(mu = 1.1e-4, lambda = 0.1, Ns = 5)

}
