% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriftFunctions.R
\name{th_PulseChase}
\alias{th_PulseChase}
\title{Analytical solution to stem cell label drift model. 
\code{th_PulseChase}  Stochastic model used in "Defining stem cell dynamics in models of intestinal tumor initiation." 
Vermeulen, Morrissey, et. al. Science 2013. The model is a stochastic birth death model with absorbing boundaries at 0 and N.}
\usage{
th_PulseChase(lambda, Ns, tau, time_points, persisting = T, Pr = 0.5,
  splitNum = 0, alpha_beta = NULL)
}
\arguments{
\item{lambda}{Stem cell replacement rate.}

\item{Ns}{Number of functional stem cells per crypt.}

\item{tau}{Time delay until drift starts.}

\item{time_points}{vector of time point values.}

\item{persisting}{Whether or not to return probability distribution of only the surviving clones or not.}

\item{Pr}{Bias in stem cell replacement Pr = 0.5 is neutral.}

\item{splitNum}{To match to data choose the number of bins to distribute the model into (e.g 1/8s, splitNum = 8).}

\item{alpha_beta}{two dim vector giving rate of replacement, alternative to using Pr. The relation is beta   = 2*lambda*Pr, alpha  = 2*lambda*(1-Pr)}
}
\value{
a matrix with the solution to the neutral or biased drift model.
}
\description{
Analytical solution to stem cell label drift model. 
\code{th_PulseChase}  Stochastic model used in "Defining stem cell dynamics in models of intestinal tumor initiation." 
Vermeulen, Morrissey, et. al. Science 2013. The model is a stochastic birth death model with absorbing boundaries at 0 and N.
}
\examples{
x = th_PulseChase(lambda = 0.1, Ns = 5, tau = 1, time_points = 1:50)
}
